+++ pkgs/test/test/runner/browser/chrome_test.dart
@@ -47,9 +47,9 @@ webSocket.addEventListener("open", function() {
   test('reports an error in onExit', () {
     var chrome = Chrome(Uri.https('dart.dev'), configuration(),
         settings: ExecutableSettings(
-            linuxExecutable: '_does_not_exist',
-            macOSExecutable: '_does_not_exist',
-            windowsExecutable: '_does_not_exist'));
+            linuxExecutables: ['_does_not_exist'],
+            macOSExecutables: ['_does_not_exist'],
+            windowsExecutables: ['_does_not_exist']));
     expect(
         chrome.onExit,
         throwsA(isApplicationException(