+++ pkgs/test/test/runner/browser/safari_test.dart
@@ -45,9 +45,9 @@ webSocket.addEventListener("open", function() {
   test('reports an error in onExit', () {
     var safari = Safari(Uri.https('dart.dev'),
         settings: ExecutableSettings(
-            linuxExecutable: '_does_not_exist',
-            macOSExecutable: '_does_not_exist',
-            windowsExecutable: '_does_not_exist'));
+            linuxExecutables: ['_does_not_exist'],
+            macOSExecutables: ['_does_not_exist'],
+            windowsExecutables: ['_does_not_exist']));
     expect(
         safari.onExit,
         throwsA(isApplicationException(