+++ pkgs/test/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 1.25.3-wip
 
 * Remove outdated StreamMatcher link from README table of contents.
-* Fix path to launch firefox on macOS.
+* Add support for the new firefox binary location on macOS.
 
 ## 1.25.2